extract safe collection traits into microcrate
