+++ Cargo.toml
@@ -25,6 +25,7 @@ members = [
     "crates/copybook-codec-memory",
     "crates/copybook-sequence-ring",
     "crates/copybook-safe-ops",
+    "crates/copybook-safe-collections",
     "crates/copybook-rdw-predicates",
     "crates/copybook-safe-index",
     "crates/copybook-safe-text",
@@ -74,6 +75,7 @@ copybook-record-io = { version = "=0.4.3", path = "crates/copybook-record-io" }
 copybook-overpunch = { version = "=0.4.3", path = "crates/copybook-overpunch" }
 copybook-determinism = { version = "=0.4.3", path = "crates/copybook-determinism" }
 copybook-safe-ops = { version = "=0.4.3", path = "crates/copybook-safe-ops" }
+copybook-safe-collections = { version = "=0.4.3", path = "crates/copybook-safe-collections" }
 copybook-safe-text = { version = "=0.4.3", path = "crates/copybook-safe-text" }
 copybook-safe-index = { version = "=0.4.3", path = "crates/copybook-safe-index" }
 copybook-rdw-predicates = { version = "=0.4.3", path = "crates/copybook-rdw-predicates" }
