+++ crates/copybook-safe-collections/README.md
@@ -1,0 +1,9 @@
+# copybook-safe-collections
+
+Single-responsibility microcrate for panic-safe collection extension traits:
+
+- `OptionExt`
+- `VecExt`
+- `SliceExt`
+
+These traits convert panicking access patterns into structured `copybook-error` results.
