+++ crates/copybook-utils/Cargo.toml
@@ -22,6 +22,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 copybook-error.workspace = true
+copybook-safe-collections.workspace = true
 copybook-safe-ops.workspace = true
 
 [dev-dependencies]
